fix: Moved the constants to separate files. Helper for JSON made general. Corrected styles
