+++ src/common/components/localStorage/constants.ts
@@ -1,0 +1,3 @@
+export default {
+  localStorageKey: 'selectedGoods',
+};
