+++ src/components/cart-header/cart-header.scss
@@ -10,9 +10,9 @@
     display: flex;
     justify-content: center;
     align-items: flex-end;
-    width: 40px;
-    height: 40px;
-    padding-bottom: 2%;
+    width: 3rem;
+    height: 2.5rem;
+    padding-bottom: 0.2em;
     background-image: url('../../common/assets/icons/shopping-bag.svg');
     background-repeat: no-repeat;
     background-position: center;
