+++ src/components/cart-header/constants.ts
@@ -1,0 +1,21 @@
+export default {
+  cart: {
+    class: 'cart',
+  },
+  cartGoodsContainer: {
+    class: 'cart__total-goods-container',
+  },
+  cartGoods: {
+    class: 'cart__total-goods',
+  },
+  cartPriceContainer: {
+    class: 'cart__total-price-container',
+    text: 'Cart total : ',
+  },
+  cartPrice: {
+    class: 'cart__total-price',
+  },
+  symbol: {
+    dollar: '$',
+  },
+};
