+++ src/components/header/constants.ts
@@ -1,0 +1,5 @@
+export default {
+  header: {
+    class: 'header',
+  },
+};
