+++ src/components/header/header.scss
@@ -5,7 +5,7 @@
   justify-content: space-between;
   align-items: center;
   width: 100%;
-  padding: 2% 5%;
+  padding: 1em 3em;
   background-color: $black;
   color: $white;
 }
