+++ src/components/logo/Logo.ts
@@ -1,20 +1,8 @@
 import { TAGS } from '../../common/helpers/constants';
+import CONSTANTS from './constants';
 import './logo.scss';
 import IDrawComponent from '../../common/interface/IDrawComponent';
 import getDOMElement from '../../common/helpers/getDOMElement';
-
-const CONSTANTS = {
-  logo: {
-    class: 'logo',
-  },
-  logoLink: {
-    class: 'logo__link',
-    text: 'Online Store',
-    atributes: {
-      href: './index.html',
-    },
-  },
-};
 
 export default class Logo implements IDrawComponent {
   public draw() {
