+++ src/components/logo/constants.ts
@@ -1,0 +1,12 @@
+export default {
+  logo: {
+    class: 'logo',
+  },
+  logoLink: {
+    class: 'logo__link',
+    text: 'Online Store',
+    atributes: {
+      href: './index.html',
+    },
+  },
+};
